Fix #118, update pointer arg types
Pointers are punned to different (incorrect) types throughout CS.  This
addresses only the most egregious violations, in particular the case
where an argument is defined as a pointer to a table and in reality it
is a pointer-to-pointer (even qualified with const when it was updating
it).  The pointer mismatches are the reason for the strange "memcpy" in
these spots.

This does some other code cleanup but only scratches the surface of what
needs to be done.

There should be a follow-on task to better structure the data within the
global, this would clean up a lot of APIs.  In particular the
CS_TableInit is a particularly ugly call with many arguments, this can
be greatly simplified with a better data structure.